initial commit for extracting ts from key
